% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longair}
\alias{longair}
\title{Airfares data (longair)}
\format{A data.frame with 1000 rows and 11 variables:
\describe{
  \item{city1}{City of boarding.}
  \item{cit2}{City of landing.}
  \item{average_fare}{Average fare.}
  \item{distance}{Distance between city of boarding and landing.}
  \item{average_weekly_passengers}{Average weekly passengers.}
  \item{market_leading_airline}{Market leading airline.}
  \item{market_share}{Market share.}
  \item{avarege_return_fare}{Average return fare}
  \item{low_price_airline}{Lower price airline.}
  \item{market_share2}{Second market share}
  \item{price}{Price of travel.}
}}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0950705118304052}
}
\usage{
longair
}
\description{
Longair data contains about quarterly average airfare and average weekly passengers
for 4177 markets in 2001 of the U.S. Department of Transportation.
The data can be seen in 'Polygonal data analysis: A new framework in symbolic data analysis' paper.
}
\keyword{datasets}
