% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.omit.polygonal.R
\name{na.omit}
\alias{na.omit}
\title{Handle Missing Values in Polygonal Objects}
\usage{
na.omit(object, ...)
}
\arguments{
\item{object}{objects of the class "\emph{polygonal}".}

\item{...}{further arguments special methods could require.}
}
\value{
polygons an object of the class "\emph{polygonal}" without missing values.
}
\description{
The function omits missing polygons.
}
\examples{
y <- psim(5, 3)
y[[1]] <- NA
na.omit(y) 
}
