% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paggreg.R
\name{paggreg}
\alias{paggreg}
\title{Polygonal data aggregation}
\usage{
paggreg(data)
}
\arguments{
\item{data}{A data frame with the first column of type factor.}
}
\value{
paggreg returns an objects of class "paggregated".
}
\description{
The function obtains symbolic data from classical data through the center and radius representation.
}
\details{
The class "aggregated" is composed by two data sets from center and range represetation. 
The first and second data set represent the center and radius, respectively.
}
\examples{
cat <- as.factor(sample(1:20, 1000, replace = TRUE))
cv <- runif(1000) #classical variable
cvc <- data.frame(category = cat, cv) 
p <- paggreg(cvc)
}
