% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pconvex.R
\name{pconvex}
\alias{pconvex}
\title{Convex verification}
\usage{
pconvex(polygon)
}
\arguments{
\item{polygon}{A matrix of dimension l x 2 where l represent number of sides polygon.}
}
\value{
A boolean.
}
\description{
Verify convexity of the polygons.
}
\examples{
x <- psim(1,4)
x <- x[[1]] 
pconvex(x) 

x <- matrix(unlist(psim(1,4)), ncol= 2)
pconvex(x) #run
}
