% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcorr.R
\name{pcorr}
\alias{pcorr}
\title{Polygonal symbolic correlation}
\usage{
pcorr(polygons)
}
\arguments{
\item{polygons}{A list of matrices of dimension l x 2, where l represent number of sides polygon.}
}
\value{
The method returns a integer.
}
\description{
Compute the symbolic polygonal empirical correlation.
}
\examples{
x = psim(10, 3) #simulate 10 polygons of 3 sides
pcorr(x)
}
