% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean.R
\name{pmean}
\alias{pmean}
\title{Polygonal empiric mean}
\usage{
pmean(polygons)
}
\arguments{
\item{polygons}{A list of matrices of dimension l x 2, where l represent number of sides polygon.}
}
\value{
The method returns a vector containing the symbolic polygonal empirical mean in first and second dimension, respectively.
}
\description{
Compute the polygonal empirical mean for polygonal variable.
}
\examples{
x <- psim(10, 10) #simulate 10 polygons of 10 sides
pmean(x)
}
