% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean_id.R
\name{pmean_id}
\alias{pmean_id}
\title{Polygonal symbolic internal mean}
\usage{
pmean_id(polygon)
}
\arguments{
\item{polygon}{a matrix representing the polygon.}
}
\value{
a polygonal empiric mean of a polygon.
}
\description{
Compute the polygonal empirical mean for only one polygonal variable.
}
\examples{
x <- psim(1, 3)
x <- x[[1]]
pmean_id(x)
}
