% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplot.R
\name{pplot}
\alias{pplot}
\title{Plot polygonal symbolic variable}
\usage{
pplot(polygon)
}
\arguments{
\item{polygon}{A list of matrices with dimension l x 2 where l represents vertices number of polygon.}
}
\description{
Prints all overlaid polygons in the display. The polygons obtained through classes.
}
\examples{
x <- psim(10, 3) #simulate 10 polygons of 3 sides
pplot(x)
}
