% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psda-deprecated.R, R/ptransform.R
\name{psda-deprecated}
\alias{psda-deprecated}
\alias{ptransform}
\title{Deprecated functions in package \pkg{psda}.}
\usage{
ptransform(data, vertices)
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. The recommended functions are paggreg followed of psymbolic.
  Help pages for deprecated functions are
  available at \code{help("ptransform-deprecated")}
}
\section{\code{ptransform}}{

For \code{ptransform}, use \code{\link{paggreg}}.
}

\keyword{internal}
