% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim.R
\name{psim}
\alias{psim}
\title{Polygonal symbolic data simulation}
\usage{
psim(objects, vertices)
}
\arguments{
\item{objects}{number of objects simulated.}

\item{vertices}{number of vertex of the polygon}
}
\value{
A list of polygons.
}
\description{
Function to create a polygonal data list.
}
\examples{
psim(2, 3)
}
