% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmi.R
\name{psmi}
\alias{psmi}
\title{Polygonal internal second moment}
\usage{
psmi(polygon)
}
\arguments{
\item{polygon}{a matrix that represents a polygonal variable.}
}
\value{
The internal variance.
}
\description{
Caltulate symbolic polygonal internal second moment for polygonal data.
}
\examples{
x <- psim(1, 3) #simulate a polygon of 3 sides
psmi(x[[1]])
}
