% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptransform.R
\name{ptransform-deprecated}
\alias{ptransform-deprecated}
\title{Polygon data transformation}
\usage{
ptransform(data, vertices)
}
\arguments{
\item{data}{A data frame with the first column of type factor.}

\item{vertices}{integer that represent number of vertices of polygon.}
}
\value{
A list of matrices of dimension l x 2, where l represent number of vertices polygon transformed.
}
\description{
It transform labeled data in polygonal datas.
}
\examples{
cat <- as.factor(sample(1:20, 1000, replace = TRUE))
cv <- runif(1000) #classical variable
cvc <- data.frame(category = cat, cv) 
p <- ptransform(cvc, 4)
}
\seealso{
\code{\link{psda-deprecated}}
}
\keyword{internal}
