% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar.R
\name{pvar}
\alias{pvar}
\title{Polygonal symbolic variance}
\usage{
pvar(polygons)
}
\arguments{
\item{polygons}{A list of matrices of dimension l x 2 where l represent number of sides polygon.}
}
\value{
The method returns a bi-dimensional vector.
}
\description{
Estime the symbolic polygonal empirical variance.
}
\examples{
x <- psim(10, 3) #simulate 10 polygons of 3 sides
pvar(x)
}
