% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvari.R
\name{pvari}
\alias{pvari}
\title{Polygonal internal variance}
\usage{
pvari(polygon)
}
\arguments{
\item{polygon}{a matrix that represents a polygonal variable.}
}
\value{
The internal variance.
}
\description{
Caltulate polygonal internal variance for polygonal data.
}
\examples{
x <- psim(1, 3) #simulate a polygons of 3 sides
pvari(x[[1]])
}
