% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spolygon.R
\name{spolygon}
\alias{spolygon}
\title{Symbolic Polygon}
\usage{
spolygon(center, radius, vertices)
}
\arguments{
\item{center}{a integer that represents the barycenter of polygon.}

\item{radius}{a integer that represents the radius of polygon.}

\item{vertices}{represents the number of vertices for the polygon.}
}
\value{
matrix that represents the polygon.
}
\description{
The function obtains a simple symbolic polygon from center and radius
representation.
}
\examples{
spolygon(2.5, 3, 5) #pentagon

}
