% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wnba2014}
\alias{wnba2014}
\title{Women National Basketball Association 2014 (WNBA 2014).}
\format{A data.frame with 4022 rows and 6 variables:
\describe{
  \item{player_id}{Identification of player.}
  \item{team_pts}{Number of points made by team.}
  \item{opp_pts}{Number of points made by opponent.}
  \item{minutes}{Minutes played.}
  \item{fgatt}{Field goal attempts.}
  \item{efficiency}{Efficiency.}
}}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0950705118304052}
}
\usage{
wnba2014
}
\description{
The data set contains information about the season 2014.
    The data can be seen in 'Polygonal data analysis: A new framework in symbolic data analysis' paper.
}
\keyword{datasets}
