% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parea.R
\name{parea}
\alias{parea}
\title{Polygonal Area}
\usage{
parea(polygon)
}
\arguments{
\item{polygon}{a matrix representing the polygon.}
}
\value{
a integer the area of polygon.
}
\description{
Compute the area of polygon.
}
\examples{
x <- psim(10, 10) #simulate 10 polygons of 10 sides
x <- x[[1]] 
parea(x)
}
