% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcov.R
\name{pcov}
\alias{pcov}
\title{Polygonal symbolic covariance}
\usage{
pcov(polygons)
}
\arguments{
\item{polygons}{A list of polygonal datas.}
}
\value{
The method returns a integer.
}
\description{
Compute the symbolic polygonal empirical covariance.
}
\examples{
x <- psim(10, 10) #simulate 10 polygons of 10 sides
pcov(x) 
}
