% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfreq.R
\name{pfreq}
\alias{pfreq}
\title{Polygonal Symbolic Relative Frequency}
\usage{
pfreq(pol)
}
\arguments{
\item{pol}{A list of matrices of dimension l x 2, where l represent number of sides polygon.}
}
\description{
Compute the bivariate relative frequency.
}
\examples{
x <- psim(10, 10) #simulate 10 polygons of 10 sides
frequency <- pfreq(x)
}
