% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean_id.R
\name{pmean_id}
\alias{pmean_id}
\title{Polygonal symbolic internal mean}
\usage{
pmean_id(polygon)
}
\arguments{
\item{polygon}{a matrix representing the polygon.}
}
\value{
a polygonal empiric mean of a polygon.
}
\description{
Compute the symbolic polygonal empirical mean for only one observation (classes).
}
\examples{
x <- psim(10, 10) #simulate 10 polygons of 10 sides
x <- x[[1]]
pmean_id(x)
}
