% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim.R
\name{psim}
\alias{psim}
\title{Polygonal symbolic data simulation}
\usage{
psim(n, vertices)
}
\arguments{
\item{n}{number of simulated polygons.}

\item{vertices}{number of vertex of the polygon.}
}
\value{
A list of polygons.
}
\description{
Simulate a polygonal variable with one or more individuals.
}
\details{
The argument \code{radius} should have all values greater than zero. Otherwise,
we cannot construct the polygons that compose the symbolic polygonal random variable. Besides,
the size of the \code{center} vector should be equal to \code{range} vector.
}
\examples{
number_polygons <- 10
psim(number_polygons, 4) 
}
