% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar.R
\name{pvar}
\alias{pvar}
\title{Polygonal symbolic variance}
\usage{
pvar(polygons)
}
\arguments{
\item{polygons}{A list of matrices of dimension l x 2 where l represent number of sides polygon.}
}
\value{
The method returns a bi-dimensional vector.
}
\description{
Estime the symbolic polygonal empirical variance.
}
\examples{
x <- psim(8, 12) #simulate 8 polygons of 12 sides
pvar(x)
}
