% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvari.R
\name{pvari}
\alias{pvari}
\title{Polygonal internal variance}
\usage{
pvari(polygon)
}
\arguments{
\item{polygon}{a matrix that represents a polygonal variable.}
}
\value{
The internal variance.
}
\description{
Caltulate the symbolic polygonal internal variance for a polygonal data.
}
\examples{
x <- psim(10, 10) #simulate 10 polygons of 10 sides
pvari(x[[1]])
}
